# openapi/jobvis-v1.yaml
openapi: 3.1.0
info:
  title: Jobvis API
  version: 1.0.0
  description: Canonical HTTP contract for Jobvis v1, including the current web client surface.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /applications:
    post:
      operationId: createApplication
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateApplicationRequest' }
      responses:
        '201':
          description: Created application.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /applications/counts:
    get:
      operationId: getApplicationCounts
      responses:
        '200':
          description: Unfiltered application count.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApplicationCounts' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /applications/page:
    get:
      operationId: listApplicationPage
      parameters:
        - $ref: '#/components/parameters/ApplicationQuery'
        - $ref: '#/components/parameters/ApplicationStatusFilter'
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200':
          description: A bounded application page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApplicationPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /applications/{applicationId}:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
    get:
      operationId: getApplication
      responses:
        '200':
          description: Core application detail; histories are fetched from their bounded endpoints.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /applications/{applicationId}/details:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
    patch:
      operationId: updateApplicationDetails
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateApplicationDetailsRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApplicationResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /applications/{applicationId}/memo:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
    put:
      operationId: updateApplicationMemo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateMemoRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApplicationResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /applications/{applicationId}/status:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
    post:
      operationId: updateApplicationStatus
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateStatusRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApplicationResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /applications/{applicationId}/schedule:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
    get:
      operationId: getApplicationSchedule
      responses:
        '200':
          description: Full all-day or timed schedule.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApplicationSchedule' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: patchApplicationSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatchScheduleRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApplicationResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    put:
      operationId: replaceApplicationSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReplaceScheduleRequest' }
      responses:
        '200':
          description: Replaced schedule.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApplicationSchedule' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /applications/{applicationId}/schedule/complete:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
    post:
      operationId: completeApplicationSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MutationRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApplicationResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /applications/{applicationId}/review/complete:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
    post:
      operationId: completeApplicationReview
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MutationRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApplicationResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /applications/review/complete-bulk:
    post:
      operationId: completeAllApplicationReviews
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompleteBulkReviewRequest' }
      responses:
        '200':
          description: Review completion result for the current review revision.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CompleteBulkReviewResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /applications/{applicationId}/activities/{activityId}:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
      - name: activityId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      operationId: deleteApplicationActivity
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MutationRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApplicationResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /applications/{applicationId}/emails:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
      - $ref: '#/components/parameters/HistoryBefore'
      - $ref: '#/components/parameters/HistoryLimit'
    get:
      operationId: listApplicationEmails
      responses:
        '200':
          description: Email history page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmailHistoryPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /applications/{applicationId}/activities:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
      - $ref: '#/components/parameters/HistoryBefore'
      - $ref: '#/components/parameters/HistoryLimit'
    get:
      operationId: listApplicationActivities
      responses:
        '200':
          description: Activity history page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActivityHistoryPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /applications/{applicationId}/changes:
    parameters:
      - $ref: '#/components/parameters/ApplicationId'
      - $ref: '#/components/parameters/HistoryBefore'
      - $ref: '#/components/parameters/HistoryLimit'
    get:
      operationId: listApplicationChanges
      responses:
        '200':
          description: Change history page.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChangeHistoryPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    ApplicationId:
      name: applicationId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ApplicationQuery:
      name: q
      in: query
      schema: { type: string }
    ApplicationStatusFilter:
      name: status
      in: query
      schema:
        $ref: '#/components/schemas/ApplicationFilter'
    HistoryBefore:
      name: before
      in: query
      schema: { type: integer, format: int64 }
    HistoryLimit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
  responses:
    ApplicationResponse:
      description: Updated application.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Application' }
    BadRequest:
      description: Invalid request.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Unauthorized:
      description: Authentication is required.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    NotFound:
      description: Owned resource was not found.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Conflict:
      description: Mutation id or optimistic version conflict.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
...
  schemas:
...
    ApplicationStatus:
      type: string
      enum: [applied, screening, test, interview, offer, offered, rejected]
    ApplicationFilter:
      type: string
      default: all
      enum: [all, review, application, test, interview, offer, offered, rejected, schedulable]
    ApplicationSourceType:
      type: string
      enum: [manual, gmail, outlook, naver, other]
    ScheduleType:
      type: string
      enum: [application, test, interview, followup, other]
    ActivityType:
      type: string
      enum: [email, note, status, task]
    MutationRequest:
      type: object
      additionalProperties: false
      required: [mutationId, expectedVersion]
      properties:
        mutationId: { type: string, format: uuid }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
    CreateApplicationRequest:
      type: object
      additionalProperties: false
      required: [mutationId, company, position, status]
      properties:
        mutationId: { type: string, format: uuid }
        company: { type: string, minLength: 1, maxLength: 160 }
        position: { type: string, minLength: 1, maxLength: 160 }
        status: { type: string, enum: [applied, screening, test, interview, offer] }
    UpdateApplicationDetailsRequest:
      type: object
      additionalProperties: false
      required: [mutationId, expectedVersion, company, position, location, employmentType, appliedAt]
      properties:
        mutationId: { type: string, format: uuid }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
        company: { type: string, minLength: 1, maxLength: 160 }
        position: { type: string, minLength: 1, maxLength: 160 }
        location: { type: [string, 'null'], maxLength: 160 }
        employmentType: { type: [string, 'null'], maxLength: 80 }
        appliedAt: { type: string, format: date }
    UpdateMemoRequest:
      type: object
      additionalProperties: false
      required: [mutationId, expectedVersion, memo]
      properties:
        mutationId: { type: string, format: uuid }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
        memo: { type: string, maxLength: 20000 }
    UpdateStatusRequest:
      type: object
      additionalProperties: false
      required: [mutationId, expectedVersion, status]
      properties:
        mutationId: { type: string, format: uuid }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
        status: { $ref: '#/components/schemas/ApplicationStatus' }
    PatchScheduleRequest:
      type: object
      additionalProperties: false
      required: [mutationId, expectedVersion]
      properties:
        mutationId: { type: string, format: uuid }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
        nextActionAt: { type: string, format: date }
        nextActionTitle: { type: string, maxLength: 200 }
      anyOf:
        - required: [nextActionAt]
        - required: [nextActionTitle]
    ReplaceScheduleRequest:
      type: object
      additionalProperties: false
      required: [mutationId, expectedVersion, scheduleType, action, scheduledAt, endsAt, timezone, location, description]
      properties:
        mutationId: { type: string, format: uuid }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
        expectedScheduleVersion: { type: integer, format: int64, minimum: 0 }
        scheduleType: { $ref: '#/components/schemas/ScheduleType' }
        action: { type: string, minLength: 1, maxLength: 200 }
        scheduledAt: { type: string, format: date-time }
        endsAt: { type: [string, 'null'], format: date-time }
        timezone: { type: string, minLength: 1, maxLength: 80 }
        location: { type: string, maxLength: 300 }
        description: { type: string, maxLength: 4000 }
    ApplicationListItem:
      type: object
      additionalProperties: false
      required: [id, version, company, position, appliedAt, status, needsReview, source]
      properties:
        id: { type: string, format: uuid }
        version: { type: integer, format: int64, minimum: 0 }
        company: { type: string }
        position: { type: string }
        appliedAt: { type: string, format: date }
        status: { $ref: '#/components/schemas/ApplicationStatus' }
        needsReview: { type: boolean }
        source: { type: string }
    Application:
      type: object
      additionalProperties: false
      required: [id, version, company, position, location, employmentType, appliedAt, status, needsReview, source, sourceType, schedule, memo]
      properties:
        id: { type: string, format: uuid }
        version: { type: integer, format: int64, minimum: 0 }
        company: { type: string }
        position: { type: string }
        location: { type: string }
        employmentType: { type: string }
        appliedAt: { type: string, format: date }
        status: { $ref: '#/components/schemas/ApplicationStatus' }
        needsReview: { type: boolean }
        source: { type: string }
        sourceType: { $ref: '#/components/schemas/ApplicationSourceType' }
        schedule:
          oneOf:
            - $ref: '#/components/schemas/ApplicationScheduleSummary'
            - type: 'null'
        memo: { type: string }
    ApplicationScheduleSummary:
      type: object
      additionalProperties: false
      required: [nextActionTitle, nextActionAt]
      properties:
        nextActionTitle: { type: string }
        nextActionAt: { type: string, format: date }
    ApplicationPage:
      type: object
      additionalProperties: false
      required: [items, page, limit, hasNext, filteredCount, totalCount, needsReviewCount, reviewRevision]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ApplicationListItem' } }
        page: { type: integer, minimum: 0 }
        limit: { type: integer, minimum: 1 }
        hasNext: { type: boolean }
        filteredCount: { type: integer, format: int64, minimum: 0 }
        totalCount: { type: integer, format: int64, minimum: 0 }
        needsReviewCount: { type: integer, format: int64, minimum: 0 }
        reviewRevision: { type: integer, format: int64, minimum: 0 }
    ApplicationCounts:
      type: object
      additionalProperties: false
      required: [totalCount]
      properties:
        totalCount: { type: integer, format: int64, minimum: 0 }
    CompleteBulkReviewRequest:
      type: object
      additionalProperties: false
      required: [mutationId, expectedReviewRevision]
      properties:
        mutationId: { type: string, format: uuid }
        expectedReviewRevision: { type: integer, format: int64, minimum: 0 }
    CompleteBulkReviewResponse:
      type: object
      additionalProperties: false
      required: [completedCount, needsReviewCount, reviewRevision]
      properties:
        completedCount: { type: integer, minimum: 0 }
        needsReviewCount: { type: integer, format: int64, minimum: 0 }
        reviewRevision: { type: integer, format: int64, minimum: 0 }
    ApplicationEmail:
      type: object
      additionalProperties: false
      required: [id, subject, sender, receivedAt, summary]
      properties:
        id: { type: string, format: uuid }
        subject: { type: string }
        sender: { type: string }
        receivedAt: { type: string, format: date-time }
        summary: { type: string }
    ApplicationActivity:
      type: object
      additionalProperties: false
      required: [id, type, title, description, occurredAt]
      properties:
        id: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/ActivityType' }
        title: { type: string }
        description: { type: string }
        occurredAt: { type: string, format: date-time }
    ApplicationChange:
      type: object
      additionalProperties: false
      required: [id, title, description, occurredAt]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        occurredAt: { type: string, format: date-time }
    ApplicationSchedule:
      oneOf:
        - $ref: '#/components/schemas/AllDayApplicationSchedule'
        - $ref: '#/components/schemas/TimedApplicationSchedule'
    AllDayApplicationSchedule:
      type: object
      additionalProperties: false
      required: [id, applicationId, applicationVersion, version, scheduleType, action, allDay, date, scheduledAt, endsAt, timezone, location, description, completed, completedAt]
      properties:
        id: { type: string, format: uuid }
        applicationId: { type: string, format: uuid }
        applicationVersion: { type: integer, format: int64, minimum: 0 }
        version: { type: integer, format: int64, minimum: 0 }
        scheduleType: { $ref: '#/components/schemas/ScheduleType' }
        action: { type: string }
        allDay: { type: boolean, const: true }
        date: { type: string, format: date }
        scheduledAt: { type: 'null' }
        endsAt: { type: 'null' }
        timezone: { type: 'null' }
        location: { type: string }
        description: { type: string }
        completed: { type: boolean }
        completedAt: { type: [string, 'null'], format: date-time }
    TimedApplicationSchedule:
      type: object
      additionalProperties: false
      required: [id, applicationId, applicationVersion, version, scheduleType, action, allDay, date, scheduledAt, endsAt, timezone, location, description, completed, completedAt]
      properties:
        id: { type: string, format: uuid }
        applicationId: { type: string, format: uuid }
        applicationVersion: { type: integer, format: int64, minimum: 0 }
        version: { type: integer, format: int64, minimum: 0 }
        scheduleType: { $ref: '#/components/schemas/ScheduleType' }
        action: { type: string }
        allDay: { type: boolean, const: false }
        date: { type: 'null' }
        scheduledAt: { type: string, format: date-time }
        endsAt: { type: [string, 'null'], format: date-time }
        timezone: { type: string }
        location: { type: string }
        description: { type: string }
        completed: { type: boolean }
        completedAt: { type: [string, 'null'], format: date-time }
    EmailHistoryPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor, totalCount]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ApplicationEmail' } }
        nextCursor: { type: [integer, 'null'], format: int64 }
        totalCount: { type: integer, format: int64, minimum: 0 }
    ActivityHistoryPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ApplicationActivity' } }
        nextCursor: { type: [integer, 'null'], format: int64 }
    ChangeHistoryPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor, totalCount]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ApplicationChange' } }
        nextCursor: { type: [integer, 'null'], format: int64 }
        totalCount: { type: integer, format: int64, minimum: 0 }
...
    ProblemDetail:
      type: object
      additionalProperties: true
      required: [title, status, detail]
      properties:
        type: { type: string, format: uri-reference, default: about:blank }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }