# src/main/kotlin/com/meenseek/jobvis/application/ApplicationController.kt
package com.meenseek.jobvis.application

import com.meenseek.jobvis.auth.CurrentUserProvider
import com.meenseek.jobvis.common.BadRequestException
import jakarta.servlet.http.HttpServletRequest
import jakarta.validation.Valid
import org.springframework.http.HttpStatus
import org.springframework.validation.annotation.Validated
import org.springframework.web.bind.annotation.DeleteMapping
import org.springframework.web.bind.annotation.GetMapping
import org.springframework.web.bind.annotation.PatchMapping
import org.springframework.web.bind.annotation.PathVariable
import org.springframework.web.bind.annotation.PostMapping
import org.springframework.web.bind.annotation.PutMapping
import org.springframework.web.bind.annotation.RequestBody
import org.springframework.web.bind.annotation.RequestMapping
import org.springframework.web.bind.annotation.RequestParam
import org.springframework.web.bind.annotation.ResponseStatus
import org.springframework.web.bind.annotation.RestController
import java.time.LocalDate
import java.time.format.DateTimeParseException
import java.util.UUID
import tools.jackson.databind.node.ObjectNode

@RestController
@Validated
@RequestMapping("/api/v1/applications")
class ApplicationController(
	private val currentUserProvider: CurrentUserProvider,
	private val applicationService: ApplicationService,
) {
	@GetMapping("/counts")
	fun counts(httpRequest: HttpServletRequest): ApplicationCountsResponse =
		applicationService.counts(currentUserProvider.currentUserId(httpRequest))

	@GetMapping("/page")
	fun listPage(
		@RequestParam(required = false) q: String?,
		@RequestParam(required = false, defaultValue = "all") status: String?,
		@RequestParam(defaultValue = "0") page: Int,
		@RequestParam(defaultValue = "50") limit: Int,
		httpRequest: HttpServletRequest,
	): ApplicationListPageResponse = applicationService.listPage(
		currentUserProvider.currentUserId(httpRequest), q, status, page, limit,
	)

	@GetMapping("/{applicationId}")
	fun get(
		@PathVariable applicationId: UUID,
		httpRequest: HttpServletRequest,
	): ApplicationResponse =
		applicationService.get(currentUserProvider.currentUserId(httpRequest), applicationId)

	@GetMapping("/{applicationId}/schedule")
	fun getSchedule(
		@PathVariable applicationId: UUID,
		httpRequest: HttpServletRequest,
	): ApplicationScheduleResponse =
		applicationService.getSchedule(currentUserProvider.currentUserId(httpRequest), applicationId)

	@GetMapping("/{applicationId}/emails")
	fun emails(
		@PathVariable applicationId: UUID,
		@RequestParam(required = false) before: Long?,
		@RequestParam(defaultValue = "50") limit: Int,
		httpRequest: HttpServletRequest,
	): HistoryPageResponse<EmailResponse> = applicationService.emails(
		currentUserProvider.currentUserId(httpRequest), applicationId, before, limit,
	)

	@GetMapping("/{applicationId}/activities")
	fun activities(
		@PathVariable applicationId: UUID,
		@RequestParam(required = false) before: Long?,
		@RequestParam(defaultValue = "50") limit: Int,
		httpRequest: HttpServletRequest,
	): HistoryPageResponse<ActivityResponse> = applicationService.activities(
		currentUserProvider.currentUserId(httpRequest), applicationId, before, limit,
	)

	@GetMapping("/{applicationId}/changes")
	fun changes(
		@PathVariable applicationId: UUID,
		@RequestParam(required = false) before: Long?,
		@RequestParam(defaultValue = "50") limit: Int,
		httpRequest: HttpServletRequest,
	): HistoryPageResponse<ChangeResponse> = applicationService.changes(
		currentUserProvider.currentUserId(httpRequest), applicationId, before, limit,
	)

	@PostMapping
	@ResponseStatus(HttpStatus.CREATED)
	fun create(
		@Valid @RequestBody request: CreateRequest,
		httpRequest: HttpServletRequest,
	): ApplicationResponse =
		applicationService.create(currentUserProvider.currentUserId(httpRequest), request)

	@PatchMapping("/{applicationId}/details")
	fun updateDetails(
		@PathVariable applicationId: UUID,
		@Valid @RequestBody request: UpdateDetailsRequest,
		httpRequest: HttpServletRequest,
	): ApplicationResponse =
		applicationService.updateDetails(currentUserProvider.currentUserId(httpRequest), applicationId, request)

	@PutMapping("/{applicationId}/memo")
	fun updateMemo(
		@PathVariable applicationId: UUID,
		@Valid @RequestBody request: UpdateMemoRequest,
		httpRequest: HttpServletRequest,
	): ApplicationResponse =
		applicationService.updateMemo(currentUserProvider.currentUserId(httpRequest), applicationId, request)

	@PostMapping("/{applicationId}/status")
	fun updateStatus(
		@PathVariable applicationId: UUID,
		@Valid @RequestBody request: UpdateStatusRequest,
		httpRequest: HttpServletRequest,
	): ApplicationResponse =
		applicationService.updateStatus(currentUserProvider.currentUserId(httpRequest), applicationId, request)

	@PostMapping("/{applicationId}/schedule/complete")
	fun completeSchedule(
		@PathVariable applicationId: UUID,
		@Valid @RequestBody request: MutationRequest,
		httpRequest: HttpServletRequest,
	): ApplicationResponse =
		applicationService.completeSchedule(currentUserProvider.currentUserId(httpRequest), applicationId, request)

	@PutMapping("/{applicationId}/schedule")
	fun updateSchedule(
		@PathVariable applicationId: UUID,
		@Valid @RequestBody request: UpdateScheduleRequest,
		httpRequest: HttpServletRequest,
	): ApplicationScheduleResponse =
		applicationService.updateSchedule(currentUserProvider.currentUserId(httpRequest), applicationId, request)

	@PatchMapping("/{applicationId}/schedule")
	fun patchSchedule(
		@PathVariable applicationId: UUID,
		@RequestBody body: ObjectNode,
		httpRequest: HttpServletRequest,
	): ApplicationResponse {
		val nextActionAtPresent = body.has("nextActionAt")
		val nextActionTitlePresent = body.has("nextActionTitle")
		if (!nextActionAtPresent && !nextActionTitlePresent) {
			throw BadRequestException("변경할 일정 날짜 또는 제목을 입력해 주세요.")
		}
		return applicationService.patchSchedule(
			currentUserProvider.currentUserId(httpRequest),
			applicationId,
			PatchScheduleRequest(
				mutationId = parseUuid(body, "mutationId"),
				expectedVersion = parseExpectedVersion(body),
				nextActionAtPresent = nextActionAtPresent,
				nextActionAt = if (nextActionAtPresent) parseDate(body, "nextActionAt") else null,
				nextActionTitlePresent = nextActionTitlePresent,
				nextActionTitle = if (nextActionTitlePresent) parseTitle(body, "nextActionTitle") else null,
			),
		)
	}

	@DeleteMapping("/{applicationId}/activities/{activityId}")
	fun deleteActivity(
		@PathVariable applicationId: UUID,
		@PathVariable activityId: UUID,
		@Valid @RequestBody request: MutationRequest,
		httpRequest: HttpServletRequest,
	): ApplicationResponse = applicationService.deleteActivity(
		currentUserProvider.currentUserId(httpRequest), applicationId, activityId, request,
	)

	@PostMapping("/{applicationId}/review/complete")
	fun completeReview(
		@PathVariable applicationId: UUID,
		@Valid @RequestBody request: MutationRequest,
		httpRequest: HttpServletRequest,
	): ApplicationResponse =
		applicationService.completeReview(currentUserProvider.currentUserId(httpRequest), applicationId, request)

	@PostMapping("/review/complete-bulk")
	fun completeBulkReview(
		@Valid @RequestBody request: CompleteBulkReviewRequest,
		httpRequest: HttpServletRequest,
	): CompleteBulkReviewResponse = applicationService.completeBulkReview(
		currentUserProvider.currentUserId(httpRequest), request,
	)

	private fun parseUuid(body: ObjectNode, field: String): UUID = try {
		val node = body.get(field)
		if (node == null || !node.isString || node.stringValue().isBlank()) throw IllegalArgumentException()
		UUID.fromString(node.stringValue())
	} catch (_: IllegalArgumentException) {
		throw BadRequestException("$field 값을 확인해 주세요.")
	}

	private fun parseExpectedVersion(body: ObjectNode): Long {
		val node = body.get("expectedVersion")
		if (node == null || !node.isIntegralNumber || !node.canConvertToLong() || node.longValue() < 0) {
			throw BadRequestException("expectedVersion 값을 확인해 주세요.")
		}
		return node.longValue()
	}

	private fun parseDate(body: ObjectNode, field: String): LocalDate = try {
		val node = body.get(field)
		if (node == null || !node.isString) throw IllegalArgumentException()
		LocalDate.parse(node.stringValue())
	} catch (_: DateTimeParseException) {
		throw BadRequestException("$field 값을 확인해 주세요.")
	} catch (_: IllegalArgumentException) {
		throw BadRequestException("$field 값을 확인해 주세요.")
	}

	private fun parseTitle(body: ObjectNode, field: String): String {
		val node = body.get(field)
		if (node == null || !node.isString) throw BadRequestException("$field 값을 확인해 주세요.")
		val value = node.stringValue().trim()
		if (value.length > 200) throw BadRequestException("$field 값을 확인해 주세요.")
		return value
	}
}